+++ R/utils.R
@@ -173,16 +173,6 @@ check_pred <- function(x, n) {
   stop("Predictions must be a length n vector or a matrix/data.frame/array with n rows.")
 }
 
-# # Is background data too small or large?
-# check_bg_size <- function(n) {
-#   if (n > 1000L) {
-#     message("Your background data 'bg_X' is large, which will slow down the process. Consider using 100-200 rows.")
-#   }
-#   if (n < 20L) {
-#     message("Your background data 'bg_X' is small, which might lead to imprecise SHAP values. Consider using 100-200 rows.")
-#   }
-# }
-
 # # Given p and maximum m, determine hybrid degree (currently not used)
 # find_degree <- function(p, m_max) {
 #   if (p < 2L) {