+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes
 
-- The background dataset is now allowed to have just a single row. This is useful in situations where there is a natural "off" value such as with MNIST data or for models that can naturally deal with missing values.
+- New argument `feature_names` allows to specify the features to calculate SHAP values for. The default equals to `colnames(X)`. This should be changed only in situations when `X` (the dataset to be explained) contains non-feature columns.
+- The background dataset can now consist of a single row only. This is useful in situations with natural "off" value such as for image data or for models that can naturally deal with missing values.
 
 
 # kernelshap 0.3.0